Seperate applyhomepage into its own app
