+++ opentech/apply/home/__init__.py
@@ -1,0 +1,1 @@
+default_app_config = 'opentech.apply.home.apps.HomeConfig'
