+++ opentech/apply/home/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class HomeConfig(AppConfig):
+    name = 'opentech.apply.home'
+    label = 'apply_home'
