+++ opentech/apply/home/models.py
@@ -1,0 +1,29 @@
+from wagtail.wagtailadmin.edit_handlers import FieldPanel
+from wagtail.wagtailcore.models import Page
+from wagtail.wagtailsearch import index
+
+from django.db import models
+
+from opentech.public.utils.models import SocialFields, ListingFields
+
+
+class ApplyHomePage(Page, SocialFields, ListingFields):  # type: ignore
+    # Only allow creating HomePages at the root level
+    parent_page_types = ['wagtailcore.Page']
+    subpage_types = ['funds.FundType']
+
+    strapline = models.CharField(blank=True, max_length=255)
+
+    search_fields = Page.search_fields + [
+        index.SearchField('strapline'),
+    ]
+
+    content_panels = Page.content_panels + [
+        FieldPanel('strapline'),
+    ]
+
+    promote_panels = (
+        Page.promote_panels +
+        SocialFields.promote_panels +
+        ListingFields.promote_panels
+    )
